Refactor : 좋아요 타입 변경
- 기존 Integer/INT 타입에서 Long/BIGINT 타입으로 변경
- 조회 관련 저장소 메서드명 구체화

+++ src/main/java/com/prgrms/kream/domain/style/dto/response/GetFeedFacadeResponse.java
@@ -7,7 +7,7 @@ public record GetFeedFacadeResponse(
 		Long id,
 		Long authorId,
 		String content,
-		Integer likes,
+		Long likes,
 		List<String> images,
 		LocalDateTime createdAt,
 		LocalDateTime updatedAt
